updated PropTypes, abstracted out Side Panes into new SidePane component
