+++ src/components/toolbar/cell-menu-subsection.jsx
@@ -6,23 +6,6 @@ import NotebookMenuHeader from './notebook-menu-header'
 
 import tasks from '../../task-definitions'
 
-// export default [
-//   <NotebookMenuItem key={tasks.moveCellUp.title} task={tasks.moveCellUp} />,
-//   <NotebookMenuItem key={tasks.moveCellDown.title} task={tasks.moveCellDown} />,
-//   <NotebookMenuItem key={tasks.addCellBelow.title} task={tasks.addCellBelow} />,
-//   <NotebookMenuItem key={tasks.deleteCell.title} task={tasks.deleteCell} />,
-//   <NotebookMenuDivider />,
-//   <NotebookMenuHeader title="change cell to ..." />,
-//   <NotebookMenuItem
-//  key={tasks.changeToJavascriptCell.title} task={tasks.changeToJavascriptCell} />,
-//   <NotebookMenuItem key={tasks.changeToMarkdownCell.title} task={tasks.changeToMarkdownCell} />,
-//   <NotebookMenuItem key={tasks.changeToCSSCell.title} task={tasks.changeToCSSCell} />,
-//   <NotebookMenuItem
-//     key={tasks.changeToExternalResourceCell.title}
-//     task={tasks.changeToExternalResourceCell}
-//   />,
-//   <NotebookMenuItem key={tasks.changeToRawCell.title} task={tasks.changeToRawCell} />]
-
 export default (
   <div>
     <NotebookMenuItem key={tasks.moveCellUp.title} task={tasks.moveCellUp} />
