+++ src/components/toolbar/declared-variables-pane.jsx
@@ -2,68 +2,24 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
-// import { ToolbarGroup } from 'material-ui/Toolbar'
-import Drawer from 'material-ui/Drawer'
-// import Close from 'material-ui/svg-icons/navigation/close'
-import { MuiThemeProvider, createMuiTheme } from 'material-ui/styles';
-import Typography from 'material-ui/Typography'
-import Close from 'material-ui-icons/Close'
-import NotebookTaskButton from './notebook-task-button'
-import NotebookMenuDivider from './notebook-menu-divider'
+import SidePane from './side-pane'
 
 import DeclaredVariables from '../declared-variables'
 import tasks from '../../task-definitions'
 
-
-const theme = createMuiTheme({
-  palette: {
-    type: 'light',
-  },
-})
-
-
 export class DeclaredVariablesPaneUnconnected extends React.Component {
-    static propTypes = {
-      sidePaneMode: PropTypes.string,
-      declaredVariables: PropTypes.object,
-    }
-
-    static muiName ='Drawer'
-
-    render() {
-      return (
-        <MuiThemeProvider theme={theme}>
-
-          <Drawer
-            classes={{ paperAnchorRight: 'side-pane' }}
-            variant="persistent"
-            anchor="right"
-            docked={false}
-            open={this.props.sidePaneMode === 'declared variables'}
-            openSecondary
-            overlayStyle={{ backgroundColor: 'none' }}
-            onRequestChange={() => { tasks.toggleDeclaredVariablesPane.callback() }}
-          >
-            <div className="pane-title">
-              <NotebookTaskButton
-                tooltip="Close"
-                task={tasks.toggleDeclaredVariablesPane}
-                style={{ color: 'black', margin: '5px' }}
-              >
-                <Close />
-              </NotebookTaskButton>
-              <Typography variant="headline">Declared Variables</Typography>
-            </div>
-            <NotebookMenuDivider />
-            <DeclaredVariables variables={this.props.declaredVariables} />
-          </Drawer>
-        </MuiThemeProvider>
-
-      )
-    }
+  static propTypes = {
+    history: PropTypes.array,
+    declaredVariables: PropTypes.object,
+  }
+  render() {
+    return (
+      <SidePane task={tasks.toggleDeclaredVariablesPane} title="Declared Variables" openOnMode="declared variables">
+        <DeclaredVariables variables={this.props.declaredVariables} />
+      </SidePane>
+    )
+  }
 }
-//            <h1 className="overlay-title">Declared Variables</h1>
-
 
 function mapStateToProps(state) {
   return {
